Factor tt::is_iterable into its own file
