+++ src/Utilities/TypeTraits/IsMaplike.hpp
@@ -7,6 +7,7 @@
 
 #include "Utilities/Requires.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/IsIterable.hpp"
 
 namespace tt {
 // @{
