+++ tests/Unit/Utilities/TypeTraits/CMakeLists.txt
@@ -15,6 +15,7 @@ set(LIBRARY_SOURCES
   Test_IsCallable.cpp
   Test_IsComplexOfFundamental.cpp
   Test_IsInteger.cpp
+  Test_IsIterable.cpp
   Test_IsMaplike.cpp
   Test_IsStreamable.cpp
   Test_RemoveReferenceWrapper.cpp
